Add search result count at export target serps

+++ scraper/src/test/java/com/serphacker/serposcope/scraper/google/scraper/GoogleScraperTest.java
@@ -194,11 +194,11 @@ public class GoogleScraperTest {
     @Test
     public void extractResults() {
         GoogleScraper scraper = new GoogleScraper(null, null);
-        assertEquals(2490l, scraper.extractResultsNumber("Environ 2 490 résultats"));
+        assertEquals(2490l, scraper.extractResultsNumber("Environ 2 490 résultats").longValue());
 //        assertEquals(25270000000l, scraper.extractResultsNumber("Page&nbsp;10 sur environ 25&nbsp;270&nbsp;000&nbsp;000&nbsp;résultats<nobr> (0,46&nbsp;secondes)&nbsp;</nobr>"));
 //        assertEquals(25270000000l, scraper.extractResultsNumber("Page 10 of about 25,270,000,000 results<nobr> (0.42 seconds)&nbsp;</nobr>"));
-        assertEquals(25270000000l, scraper.extractResultsNumber("About 25,270,000,000 results<nobr> (0.28 seconds)&nbsp;</nobr>"));
-        assertEquals(225000l, scraper.extractResultsNumber("About 225,000 results<nobr> (0.87 seconds)&nbsp;</nobr>"));
+        assertEquals(25270000000l, scraper.extractResultsNumber("About 25,270,000,000 results<nobr> (0.28 seconds)&nbsp;</nobr>").longValue());
+        assertEquals(225000l, scraper.extractResultsNumber("About 225,000 results<nobr> (0.87 seconds)&nbsp;</nobr>").longValue());
 //        assertEquals(225000l, scraper.extractResultsNumber("Page 5 of about 225,000 results (0.45 seconds) "));
     }
